requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
aiofiles==23.2.1
pymongo==4.10.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]
pytesseract
pdf2image
pytest
httpx
fpdf2
slowapi
python-magic
# Note: For python-magic on Windows, python-magic-bin is usually needed instead. Let's add that.
python-magic-bin; sys_platform == 'win32'
